+++ source/MP4.BXML.cpp
@@ -45,7 +45,7 @@ std::string BXML::getContent( void )
     return o.str();
 }
 
-void BXML::processData( MP4::BinaryStream * stream, size_t length )
+void BXML::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
