+++ source/MP4.DINF.cpp
@@ -18,7 +18,7 @@ std::string DINF::getContent( void )
     return o.str();
 }
 
-void DINF::processData( MP4::BinaryStream * stream, size_t length )
+void DINF::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     (void) stream;
     (void) length;
