+++ source/MP4.FRMA.cpp
@@ -45,7 +45,7 @@ std::string FRMA::getContent( void )
     return o.str();
 }
 
-void FRMA::processData( MP4::BinaryStream * stream, size_t length )
+void FRMA::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
