+++ source/MP4.IMIF.cpp
@@ -45,7 +45,7 @@ std::string IMIF::getContent( void )
     return o.str();
 }
 
-void IMIF::processData( MP4::BinaryStream * stream, size_t length )
+void IMIF::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
