+++ source/MP4.IPRO.cpp
@@ -18,7 +18,7 @@ std::string IPRO::getContent( void )
     return o.str();
 }
 
-void IPRO::processData( MP4::BinaryStream * stream, size_t length )
+void IPRO::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     (void) stream;
     (void) length;
