+++ source/MP4.MFRO.cpp
@@ -45,7 +45,7 @@ std::string MFRO::getContent( void )
     return o.str();
 }
 
-void MFRO::processData( MP4::BinaryStream * stream, size_t length )
+void MFRO::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
