+++ source/MP4.SCHM.cpp
@@ -45,7 +45,7 @@ std::string SCHM::getContent( void )
     return o.str();
 }
 
-void SCHM::processData( MP4::BinaryStream * stream, size_t length )
+void SCHM::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
