+++ source/MP4.SDTP.cpp
@@ -45,7 +45,7 @@ std::string SDTP::getContent( void )
     return o.str();
 }
 
-void SDTP::processData( MP4::BinaryStream * stream, size_t length )
+void SDTP::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
