+++ source/MP4.SKIP.cpp
@@ -18,7 +18,7 @@ std::string SKIP::getContent( void )
     return o.str();
 }
 
-void SKIP::processData( MP4::BinaryStream * stream, size_t length )
+void SKIP::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     (void) stream;
     (void) length;
