+++ source/MP4.STBL.cpp
@@ -18,7 +18,7 @@ std::string STBL::getContent( void )
     return o.str();
 }
 
-void STBL::processData( MP4::BinaryStream * stream, size_t length )
+void STBL::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     (void) stream;
     (void) length;
