+++ source/MP4.STTS.cpp
@@ -53,7 +53,7 @@ std::string STTS::getContent( void )
     return o.str();
 }
 
-void STTS::processData( MP4::BinaryStream * stream, size_t length )
+void STTS::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     FullBox::processData(stream, length);
