+++ source/MP4.TRAK.cpp
@@ -18,7 +18,7 @@ std::string TRAK::getContent( void )
     return o.str();
 }
 
-void TRAK::processData( MP4::BinaryStream * stream, size_t length )
+void TRAK::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     (void) stream;
     (void) length;
