+++ source/include/MP4.AVCC.h
@@ -47,7 +47,7 @@ namespace MP4
         
         std::string getName() { return "AVC Configuration"; };
         std::string getContent( void );
-        void processData( MP4::BinaryStream * stream, size_t length );
+        void processData( MP4::BinaryStream * stream, uint64_t length );
     };
 }
