+++ source/include/MP4.SBGP.h
@@ -52,7 +52,7 @@ namespace MP4
         
             std::string getName() { return "Sample to Group"; }
             std::string getContent( void );
-            void processData( MP4::BinaryStream * stream, size_t length );
+            void processData( MP4::BinaryStream * stream, uint64_t length );
     };
 }
