+++ source/include/MP4.TREX.h
@@ -52,7 +52,7 @@ namespace MP4
         
             std::string getName() { return "Track Extends Defaults"; }
             std::string getContent( void );
-            void processData( MP4::BinaryStream * stream, size_t length );
+            void processData( MP4::BinaryStream * stream, uint64_t length );
     };
 }
