+++ source/include/MP4.URL.h
@@ -29,7 +29,7 @@ namespace MP4
         
         std::string getName() { return "Data Entry URL"; }
         std::string getContent( void );
-        void processData( MP4::BinaryStream * stream, size_t length );
+        void processData( MP4::BinaryStream * stream, uint64_t length );
     };
 }
